scripts/requirements.txt
httpx
beautifulsoup4
supabase
python-dotenv
tenacity
pydantic
